requirements.txt
python-dotenv===0.21.1
Django===5.0.2
djangorestframework===3.14.0
black===24.1.1

Scrapy===2.11.0
scrapy_djangoitem==1.1.1

django-tree-queries==0.15.0